+++ msproteomics/sitereport.py
@@ -1,5 +1,5 @@
 # coding=utf-8
-# Copyright 2023 Thang V Pham
+# Copyright 2025 Thang V Pham
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.